refactor: 관리자 인가 예외를 공통 경계로 이동
