+++ src/main/java/kr/kro/airbob/common/exception/AdminAccessDeniedException.java
@@ -1,7 +1,4 @@
-package kr.kro.airbob.domain.auth.exception;
-
-import kr.kro.airbob.common.exception.BaseException;
-import kr.kro.airbob.common.exception.ErrorCode;
+package kr.kro.airbob.common.exception;
 
 public class AdminAccessDeniedException extends BaseException {
