+++ src/main/java/kr/kro/airbob/domain/member/service/MemberAdminService.java
@@ -10,9 +10,9 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
+import kr.kro.airbob.common.exception.AdminAccessDeniedException;
 import kr.kro.airbob.common.history.ChangeType;
 import kr.kro.airbob.common.history.HistoryConstants;
-import kr.kro.airbob.domain.auth.exception.AdminAccessDeniedException;
 import kr.kro.airbob.domain.member.common.MemberRole;
 import kr.kro.airbob.domain.member.dto.MemberAdminRequest.ChangeRole;
 import kr.kro.airbob.domain.member.dto.MemberAdminResponse.RoleChanged;
